+++ R/mod_global.R
@@ -121,10 +121,17 @@ mod_global_server <- function(input, output, session, orig_data, orig_data_aggre
 
   # > line plot top 5
   df_top_n <- reactive({
-    world_top_5_confirmed() %>%
-      mutate(status = as.factor(Country.Region)) %>%
+    # create factors with first top confirmed
+    countries_order =  world_top_5_confirmed() %>% filter(date == max(date)) %>%
+      arrange(desc(confirmed)) %>%
+      #arrange(!!as.symbol(input$radio_indicator)) %>%
+       .[,"Country.Region"] %>% as.vector()
+
+    df_top_n = world_top_5_confirmed() %>%
+      mutate(status = factor(Country.Region, levels = countries_order[, "Country.Region", drop = T])) %>%
       mutate(value = confirmed) %>%
       capitalize_names_df()
+    df_top_n
   })
 
   callModule(mod_plot_log_linear_server, "plot_log_linear_top_n", df = df_top_n, type = "line")