+++ man-roxygen/ex-mod_compare_nth_cases_plot.R
@@ -4,6 +4,10 @@ if (interactive()) {
   library(dplyr)
   library(tidyr)
   library(shinycssloaders)
+  library(plotly)
+  library(ggplot2)
+  library(scales)
+  sapply(file.path("R",list.files("R")), source)
   ui <- fluidPage(
     tagList(
       Covid19:::golem_add_external_resources(),
@@ -11,20 +15,23 @@ if (interactive()) {
     )
   )
   server <- function(input, output, session) {
-    orig_data <- reactive({
-      get_timeseries_full_data() %>%
+      orig_data =  get_timeseries_full_data() %>%
         get_timeseries_by_contagion_day_data()
-    })
+
 
     orig_data_aggregate <- reactive({
-      orig_data_aggregate <- orig_data() %>%
+      orig_data_aggregate <- orig_data %>%
         aggregate_province_timeseries_data() %>%
         add_growth_death_rate() %>%
         arrange(Country.Region)
       orig_data_aggregate
     })
-    #TODO: colors are not fixed yet
-    callModule(mod_compare_nth_cases_plot_server, "plot_compare_nth", orig_data_aggregate)
+    n = 1000; w = 7
+    data_filtered <- reactive({
+      orig_data_aggregate() %>%
+        rescale_df_contagion(n = n, w = w)
+    })
+    callModule(mod_compare_nth_cases_plot_server, "plot_compare_nth", data_filtered)
   }
   runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
 }
@@ -56,7 +63,6 @@ if (interactive()) {
         arrange(Country.Region)
       orig_data_aggregate
     })
-    #TODO: colors are not fixed yet
     n = 1000
     countries = c("Switzerland", "Italy", "France")
     countries_data <- reactive({